+++ doc/WhatsNew.rst
@@ -7,6 +7,8 @@ Ver 1.0.1 (2025-08-27)
 - updated Subaru/MOIRCS FOV dimensions
 - Fixed a bug where Visibility could not be started without access to
   TelescopePosition
+- Added a feature to InsFov where it rotates the image by the position
+  angle to keep the orientation of the FOV overlay
 
 Ver 1.0.0 (2025-08-13)
 ======================